모듈화 끝!

+++ front/src/api/commonApi.ts
@@ -13,15 +13,16 @@ const apiClient = axios.create({
   baseURL: Config.API_APP_KEY,
 });
 
-const addHeader = (token: string) => {
-  apiClient.interceptors.request.use(config => {
-    config.headers.Authorization = `Bearer ${token}`;
-    return config;
-  });
-};
-
 //? 회원가입, 로그인 관련 api 모두 요청 가능
 const userAPI = {
+  get: async (url: string, token: string) => {
+    try {
+      const res = await apiClient.get(url, {headers: {Authorization: `Bearer ${token}`}});
+      return res;
+    } catch (error: any) {
+      throw error as AxiosError;
+    }
+  },
   post: async (url: string, data?: KeyData, option?: KeyData) => {
     try {
       const res = await apiClient.post(url, data, option);
@@ -30,18 +31,10 @@ const userAPI = {
       throw error as AxiosError;
     }
   },
-  get: async (url: string) => {
+
+  patch: async (url: string, token: string, data: KeyData) => {
     try {
-      const res = await apiClient.get(url);
-      return res;
-    } catch (error: any) {
-      throw error as AxiosError;
-    }
-  },
-  patch: async (url: string, data: KeyData, token: string) => {
-    try {
-      addHeader(token);
-      const res = await apiClient.patch(url, data);
+      const res = await apiClient.patch(url, data, {headers: {Authorization: `Bearer ${token}`}});
       return res;
     } catch (error: any) {
       throw error as AxiosError;
@@ -49,8 +42,7 @@ const userAPI = {
   },
   delete: async (url: string, token: string) => {
     try {
-      addHeader(token);
-      const res = await apiClient.delete(url);
+      const res = await apiClient.delete(url, {headers: {Authorization: `Bearer ${token}`}});
       return res;
     } catch (error: any) {
       throw error as AxiosError;
@@ -63,8 +55,9 @@ const userAPI = {
 const scheduleAPI = {
   scheduleGet: async (token: string) => {
     try {
-      addHeader(token);
-      const response = await apiClient.get('schedule');
+      const response = await apiClient.get('schedule', {
+        headers: {Authorization: `Bearer ${token}`},
+      });
       const fetchedData = response.data.body.data;
       return convertScheduleData(fetchedData);
     } catch (error) {
@@ -73,8 +66,9 @@ const scheduleAPI = {
   },
   schedulePost: async (token: string, data: SetScheduleData) => {
     try {
-      addHeader(token);
-      const response = await apiClient.post(`schedule`, data);
+      const response = await apiClient.post(`schedule`, data, {
+        headers: {Authorization: `Bearer ${token}`},
+      });
       return response.data;
     } catch (error) {
       throw error as AxiosError;
@@ -82,7 +76,9 @@ const scheduleAPI = {
   },
   schedulePatch: async (token: string, scheduleId: number, data: any) => {
     try {
-      const response = await apiClient.patch(`schedule/${scheduleId}`, data);
+      const response = await apiClient.patch(`schedule/${scheduleId}`, data, {
+        headers: {Authorization: `Bearer ${token}`},
+      });
       return response.data;
     } catch (error) {
       throw error as AxiosError;
@@ -90,7 +86,9 @@ const scheduleAPI = {
   },
   scheduleDelete: async (token: string, scheduleId: number) => {
     try {
-      const response = await apiClient.delete(`schedule/${scheduleId}`);
+      const response = await apiClient.delete(`schedule/${scheduleId}`, {
+        headers: {Authorization: `Bearer ${token}`},
+      });
       if (response.status === 200 || response.status === 204) {
         return true;
       } else {
@@ -103,7 +101,9 @@ const scheduleAPI = {
   },
   tagIdScheduleGet: async (token: string, tagId: number) => {
     try {
-      const response = await apiClient.get(`tags/${tagId}`);
+      const response = await apiClient.get(`tags/${tagId}`, {
+        headers: {Authorization: `Bearer ${token}`},
+      });
       const fetchedData = response.data.body.data.schedules;
       return convertTagData(fetchedData);
     } catch (error) {
@@ -112,7 +112,7 @@ const scheduleAPI = {
   },
   tagGet: async (token: string) => {
     try {
-      const response = await apiClient.get(`tags`);
+      const response = await apiClient.get(`tags`, {headers: {Authorization: `Bearer ${token}`}});
       return response.data.body.data;
     } catch (error) {
       throw error as AxiosError;
@@ -120,7 +120,9 @@ const scheduleAPI = {
   },
   tagPost: async (token: string, tag: SetBoardData) => {
     try {
-      const response = await apiClient.post(`tags`, tag);
+      const response = await apiClient.post(`tags`, tag, {
+        headers: {Authorization: `Bearer ${token}`},
+      });
       return response.data;
     } catch (error) {
       throw error as AxiosError;
@@ -128,7 +130,9 @@ const scheduleAPI = {
   },
   tagPatch: async (token: string, tagId: number, boardData: SetBoardData) => {
     try {
-      const response = await apiClient.patch(`tags/${tagId}`, boardData);
+      const response = await apiClient.patch(`tags/${tagId}`, boardData, {
+        headers: {Authorization: `Bearer ${token}`},
+      });
       return response.data;
     } catch (error) {
       throw error as AxiosError;
@@ -136,7 +140,9 @@ const scheduleAPI = {
   },
   tagdelete: async (token: string, tagId: number) => {
     try {
-      const response = await apiClient.delete(`tags/${tagId}`);
+      const response = await apiClient.delete(`tags/${tagId}`, {
+        headers: {Authorization: `Bearer ${token}`},
+      });
       if (response.status === 200 || response.status === 204) {
         return true;
       } else {
@@ -147,9 +153,11 @@ const scheduleAPI = {
       return false;
     }
   },
-  notificationGet: async () => {
+  notificationGet: async (token: string) => {
     try {
-      const res = await apiClient.get('notification');
+      const res = await apiClient.get('notification', {
+        headers: {Authorization: `Bearer ${token}`},
+      });
       const resData = res.data.body.data;
       if (res.status === 200) {
         return resData.sort();
@@ -158,9 +166,10 @@ const scheduleAPI = {
       throw error as AxiosError;
     }
   },
-  notificationDelete: async (notificationIds: number[]) => {
+  notificationDelete: async (token: string, notificationIds: number[]) => {
     try {
       const res = await apiClient.delete('notification', {
+        headers: {Authorization: `Bearer ${token}`},
         data: {notificationIds},
       });
       return res.status;
@@ -168,9 +177,12 @@ const scheduleAPI = {
       throw error as AxiosError;
     }
   },
-  searchGet: async (params: KeyData) => {
+  searchGet: async (token: string, params: KeyData) => {
     try {
-      const res = await apiClient.get('schedule/search', {params});
+      const res = await apiClient.get('schedule/search', {
+        headers: {Authorization: `Bearer ${token}`},
+        ...params,
+      });
       const fetchedData = res.data.body.data;
       return convertScheduleData(fetchedData);
     } catch (error) {
